FTP업로드 개발중 서치바 개발중   - 멀티토글버튼 추가완료   - 카테고리 추가완료

+++ loggalApp/app/src/main/java/com/altsoft/model/device/AD_DEVICE_MOBILE_LIST.java
@@ -1,7 +1,7 @@
 package com.altsoft.model.device;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
+import java.math.Integer;
 
 public class AD_DEVICE_MOBILE_LIST  implements Serializable {
     /// 로컬박스코드
@@ -37,7 +37,7 @@ public class AD_DEVICE_MOBILE_LIST  implements Serializable {
     /// 클릭수
     public Integer CLICK_CNT ;
     /// 별포인트 5점 만점 평점으로 계산
-    public BigDecimal GRADE_POINT ;
+    public Integer GRADE_POINT ;
     /// 요청한사용자코드 T_MEMBER 테이블의 MEMBER_CODE
     public int MEMBER_CODE ;
     /// 요청한사용자명
